+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Notable changes to WatchWare MP3 will be documented in this file.
 ## [1.4.0] - 2025-05-13
 - Jump volume by 5% instead of 2% since the OS does not support lower increments anyway
 - Player: Fix volume to react to volume changes via OS/Bluetooth
+- Player: Time labels are now clickable - left jumps backward 1 minute, right jumps forward 1 minute
 
 ## [1.3.0] - 2025-05-10